+++ src/main/java/org/tinspin/index/PointMultimap.java
@@ -23,8 +23,8 @@ import org.tinspin.index.phtree.PHTreeMMP;
 import org.tinspin.index.qthypercube.QuadTreeKD;
 import org.tinspin.index.qthypercube2.QuadTreeKD2;
 import org.tinspin.index.qtplain.QuadTreeKD0;
-import org.tinspin.index.rtree.Entry;
 import org.tinspin.index.rtree.RTree;
+import org.tinspin.index.rtree.RTreeEntry;
 import org.tinspin.index.util.PointMultimapWrapper;
 
 import java.util.Iterator;
@@ -155,14 +155,25 @@ public interface PointMultimap<T> extends Index {
         }
 
         /**
-         * Create a kD-Tree
+         * Create a kD-Tree.
          *
          * @param dims Number of dimensions.
          * @param <T>  Value type
-         * @return New PH-Tree
+         * @return New kD-Tree
          */
         static <T> PointMultimap<T> createKdTree(int dims) {
             return KDTree.create(dims);
+        }
+
+        /**
+         * Create a kD-Tree.
+         *
+         * @param cfg Index configuration.
+         * @param <T> Value type
+         * @return New kD-Tree
+         */
+        static <T> PointMultimap<T> createKdTree(IndexConfig cfg) {
+            return KDTree.create(cfg);
         }
 
         /**
@@ -258,7 +269,7 @@ public interface PointMultimap<T> extends Index {
         }
 
         /**
-         * Create an R*Tree. R*Tree can be "turned into" STR-Trees by using {@link RTree#load(Entry[])}.
+         * Create an R*Tree.
          *
          * @param dims Number of dimensions.
          * @param <T>  Value type
@@ -266,6 +277,20 @@ public interface PointMultimap<T> extends Index {
          */
         static <T> PointMultimap<T> createRStarTree(int dims) {
             return PointMultimapWrapper.create(RTree.createRStar(dims));
+        }
+
+        /**
+         * Create an STR-loaded R*Tree.
+         *
+         * @param dims    Number of dimensions.
+         * @param entries All entries of the tree. Entries can be created with
+         *                {@link RTreeEntry#createPoint(double[], Object)}
+         * @return New STR-loaded R*Tree
+         */
+        static <T> PointMultimap<T> createAndLoadStrRTree(int dims, RTreeEntry<T>[] entries) {
+            RTree<T> tree = RTree.createRStar(dims);
+            tree.load(entries);
+            return PointMultimapWrapper.create(tree);
         }
     }
 }