+++ src/test/java/org/tinspin/index/test/util/BoxMapCandidate.java
@@ -16,7 +16,7 @@ import org.tinspin.index.array.RectArray;
 import org.tinspin.index.phtree.PHTreeR;
 import org.tinspin.index.qthypercube.QuadTreeRKD;
 import org.tinspin.index.qtplain.QuadTreeRKD0;
-import org.tinspin.index.rtree.Entry;
+import org.tinspin.index.rtree.RTreeEntry;
 import org.tinspin.index.rtree.RTree;
 
 import static org.tinspin.index.Index.*;
@@ -69,7 +69,7 @@ public class BoxMapCandidate extends Candidate {
 	public void load(double[] data, int dims) {
 		this.data = data;
 		if (bulkloadSTR) {
-			Entry<Object>[] entries = new Entry[N];
+			RTreeEntry<Object>[] entries = new RTreeEntry[N];
 			int pos = 0;
 			for (int i = 0; i < N; i++) {
 				double[] lo = new double[dims];
@@ -78,7 +78,7 @@ public class BoxMapCandidate extends Candidate {
 				pos += dims;
 				System.arraycopy(data, pos, hi, 0, dims);
 				pos += dims;
-				entries[i] = new Entry<Object>(lo, hi, O);
+				entries[i] = new RTreeEntry<Object>(lo, hi, O);
 			}
 			RTree<Object> rt = (RTree<Object>) idx;
 			rt.load(entries);